+++ tests/testthat/test-spruce.R
@@ -51,3 +51,43 @@ test_that("spruce - prob", {
     "2"
   )
 })
+
+
+test_that("spruce - numeric - multi", {
+  spruced <- spruce_numeric_multi(tibble(a = 1:5, b = 2:6))
+
+  expect_s3_class(spruced, "tbl_df")
+  expect_equal(colnames(spruced), c(".pred_1",".pred_2"))
+
+  expect_error(spruce_numeric_multi(letters[1:5]))
+  expect_error(spruce_numeric_multi(tibble(a = letters[1:5])))
+  expect_error(spruce_numeric_multi(matrix(rnorm(4),nrow = 2, ncol = 2)))
+})
+
+test_that("spruce - class - multi", {
+  spruced <- spruce_class_multi(
+    tibble(a = factor(letters[1:5]), b = factor(letters[2:6]))
+  )
+
+  expect_s3_class(spruced, "tbl_df")
+  expect_equal(colnames(spruced), c(".pred_class_1", ".pred_class_2"))
+
+  expect_error(spruce_class_multi(tibble(a = 1:5)))
+  expect_error(spruce_class_multi("hi"))
+})
+
+test_that("spruce - prob - multi", {
+  pred_levels_df <- tibble(day = letters[1:5],
+                           month = letters[2:13])
+  prob_matrix_df <- tibble(day = tibble(matrix(seq(.1, .5, .1), nrow = 2, ncol = 5, byrow = T ), .name_repair = "minimal"),
+                           month = tibble(matrix(seq(.5, 1, .05), nrow = 2, ncol = 12, byrow = T), .name_repair = "minimal"))
+  spruced <- spruce_prob_multi(pred_levels_df, prob_matrix_df)
+
+  expect_s3_class(spruced, "tbl_df")
+  expect_equal(colnames(spruced), c(".pred_class_1", ".pred_class_2"))
+
+  expect_error(spruce_class_multi(tibble(a = 1:5)))
+  expect_error(spruce_class_multi("hi"))
+})
+
+