+++ NEWS.md
@@ -1,5 +1,8 @@
 # hardhat (development version)
 
+* Adds `get_indicators()` function that takes a factor as input, and returns
+  the indicators as a integer matrix.
+
 * Added more documentation about importance and frequency weights in
   `?importance_weights()` and `?frequency_weights()` (#214).